% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliations.R
\docType{data}
\name{affiliations}
\alias{affiliations}
\title{affiliations (data)}
\format{{
\describe{
\item{\code{paperid}}{numeric}
\item{\code{affiliationid}}{numeric}
\item{\code{grid_id}}{numeric}
\item{\code{authorid}}{numeric}
\item{\code{lat}}{numeric}
\item{\code{lng}}{numeric}
\item{\code{city}}{character}
\item{\code{country}}{character}
\item{\code{normalizedname}}{character}
}
}}
\usage{
data("affiliations")
}
\description{
A joint table containing the affiliation data for MAG (for Kenyan organisations) and the Lens (other publications referencing Kenya)
}
\keyword{datasets}
