% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authors.R
\docType{data}
\name{authors}
\alias{authors}
\title{authors data}
\format{{
\describe{
\item{\code{authorid}}{numeric}
\item{\code{affiliationid}}{numeric}
\item{\code{authorsequencenumber}}{numeric}
\item{\code{originalaffiliation}}{character}
\item{\code{authors_rankno}}{numeric}
\item{\code{authors_normalizedname}}{character}
\item{\code{authors_displayname}}{character}
\item{\code{authors_lastknownaffiliationid}}{numeric}
\item{\code{authors_papercount}}{numeric}
\item{\code{authors_citationcount}}{numeric}
\item{\code{authors_createdate}}{numeric}
\item{\code{source}}{numeric}
\item{\code{kenya_personname}}{logical}
}
}}
\usage{
data("authors")
}
\description{
Authors table containing author names, paperid, authorid and original affiliations
}
\keyword{datasets}
