% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fos.R
\docType{data}
\name{fos}
\alias{fos}
\alias{fos_kenya}
\title{Fields of Study for Publications From Kenya (MAG)}
\format{{
\describe{
\item{\code{paperid}}{character}
\item{\code{fieldofstudyid}}{character}
\item{\code{score}}{character}
\item{\code{rankno}}{character}
\item{\code{normalizedname}}{character}
\item{\code{displayname}}{character}
\item{\code{maintype}}{character}
\item{\code{main}}{character}
\item{\code{type}}{character}
\item{\code{level}}{character}
\item{\code{papercount}}{character}
\item{\code{citationcount}}{character}
\item{\code{createddate}}{character}
\item{\code{abs_maintype}}{character}
\item{\code{white_keep}}{character}
\item{\code{gni}}{character}
\item{\code{abs_fieldofstudy}}{character}
\item{\code{abs_displayname}}{character}
\item{\code{abs_combined}}{character}
}
}}
\usage{
data("fos")
}
\description{
Fields of Study linked to publications by Kenyan organisations appearing in Microsoft Academic Graph
}
\details{
Data in this table only refers to fields of study for publications linked to organisations in Kenya. For the wider fields of study see lens_fos.
}
\keyword{datasets}
