% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens_fos.R
\docType{data}
\name{lens_fos}
\alias{lens_fos}
\alias{for_kenya}
\title{Fields of Study for Publications about Kenya (Lens data)}
\format{{
\describe{
\item{\code{lens_id}}{character}
\item{\code{fields_of_study}}{character}
\item{\code{doi}}{character}
\item{\code{pmid}}{character}
\item{\code{pmcid}}{character}
\item{\code{microsoft_academic_id}}{character}
\item{\code{paperid}}{character}
\item{\code{in_kenya_papers}}{character}
}
}}
\usage{
data("fos")
}
\description{
Fields of Study linked to publications mentioning Kenya from the Lens
}
\details{
Data in this table only refers to fields of study for publications that are mainly from outside Kenya. For publications only from Kenyan organisations see fos.
}
\keyword{datasets}
