% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papers.R
\docType{data}
\name{papers}
\alias{papers}
\title{papers (data)}
\format{{
\describe{
\item{\code{affiliationid}}{character}
\item{\code{rankno}}{character}
\item{\code{normalizedname}}{character}
\item{\code{displayname}}{character}
\item{\code{grid_id}}{character}
\item{\code{officialpage}}{character}
\item{\code{wikipage}}{character}
\item{\code{papercount}}{character}
\item{\code{organisation_citation}}{character}
\item{\code{name}}{character}
\item{\code{wikipedia_url}}{character}
\item{\code{established}}{character}
\item{\code{lat}}{character}
\item{\code{lng}}{character}
\item{\code{postcode}}{character}
\item{\code{primary}}{character}
\item{\code{city}}{character}
\item{\code{state}}{character}
\item{\code{state_code}}{character}
\item{\code{country}}{character}
\item{\code{country_code}}{character}
\item{\code{geonames_city_id}}{character}
\item{\code{paperid}}{character}
\item{\code{authorid}}{character}
\item{\code{authorsequencenumber}}{character}
\item{\code{originalaffiliation}}{character}
\item{\code{doi}}{character}
\item{\code{doctype}}{character}
\item{\code{papertitle}}{character}
\item{\code{originaltitle}}{character}
\item{\code{booktitle}}{character}
\item{\code{year}}{character}
\item{\code{date}}{character}
\item{\code{publisher}}{character}
\item{\code{journalid}}{character}
\item{\code{conferenceseriesid}}{character}
\item{\code{conferanceinstanceid}}{character}
\item{\code{volume}}{character}
\item{\code{issue}}{character}
\item{\code{firstpage}}{character}
\item{\code{lastpage}}{character}
\item{\code{referencecount}}{character}
\item{\code{paper_citation}}{character}
\item{\code{estimatedcitation}}{character}
\item{\code{originalvenue}}{character}
\item{\code{duplicated_paperid}}{character}
}
}}
\usage{
data("papers")
}
\description{
Microsoft Academic Graph papers table
}
\keyword{datasets}
