% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacked_bar.R
\name{stacked_bar}
\alias{stacked_bar}
\title{Create a Stacked Bar Chart}
\usage{
stacked_bar(x, col = NULL, top = NULL, title = NULL,
  x_label = NULL, y_label = NULL)
}
\arguments{
\item{x}{a data frame}

\item{col}{the column to be stacked, quoted}

\item{top}{the number of items to include, numeric}

\item{title}{chart title, quoted}

\item{x_label}{chart x axis label, quoted}

\item{y_label}{chart y axis label, quoted}
}
\value{
plot
}
\description{
Draw a vertically stacked bar chart with ggplot2
}
\examples{
\dontrun{stacked_bar(affiliations, col = "normalizedname", top = 10, title = "Papers by Affiliation", x_label = "Affiliation", y_label = "Publication Count")}
}
