% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends.R
\name{trends}
\alias{trends}
\title{Create a line graph with trends}
\usage{
trends(df, x = NULL, y = NULL, start = NULL, end = NULL,
  weight = NULL)
}
\arguments{
\item{df}{a data.frame as input}

\item{x}{the column for the x axis}

\item{y}{the value for the y axis (use n if count is generated by the function)}

\item{start}{start year (numeric)}

\item{end}{end year (numeric)}

\item{weight}{The column to use as the weight (e.g. organisation, country etc)}
}
\value{
plot
}
\description{
a line plot
}
\examples{
\dontrun{trends(lens, x = "publication_year", y = n, start = 1990, end = 2016)}
}
